requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==1.2.0
Django==4.1.7
djangorestframework==3.14.0
docopt==0.6.2
duckdb==0.7.1
gunicorn==20.1.0
idna==3.4
numpy==1.24.2
pandas==1.5.3
pipreqs==0.4.11
psycopg2-binary==2.9.5
pyarrow==11.0.0
pylance==0.3.18
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
whitenoise==6.4.0
yarg==0.1.9
